chore: move button components to dedicated folder
