+++ src/components/buttons/PreviewButton.tsx
@@ -13,27 +13,20 @@ import {
 } from 'react-admin';
 import PreviewIcon from '@mui/icons-material/Preview';
 import CloseIcon from '@mui/icons-material/Close';
-
-import {
+import React, {
     Fragment,
     ReactElement,
     useCallback,
     useEffect,
     useState,
 } from 'react';
-
-import { get } from 'lodash';
-
 import AceEditor from 'react-ace';
 import 'ace-builds/src-noconflict/mode-yaml';
 import 'ace-builds/src-noconflict/mode-json';
 import 'ace-builds/src-noconflict/mode-markdown';
-
 import { LazyLog } from '@melloware/react-logviewer';
 import { usePapaParse } from 'react-papaparse';
-
-import { DataGrid, GridColDef } from '@mui/x-data-grid';
-
+import { DataGrid } from '@mui/x-data-grid';
 import {
     Box,
     Breakpoint,
@@ -48,7 +41,6 @@ import {
     Typography,
 } from '@mui/material';
 import { CreateInDialogButtonClasses } from '@dslab/ra-dialog-crud';
-import React from 'react';
 
 const defaultIcon = <PreviewIcon />;
 
@@ -85,6 +77,7 @@ export const PreviewButton = (props: PreviewButtonProps) => {
         e.stopPropagation();
         setOpen(false);
     };
+
     const handleClick = useCallback(e => {
         e.stopPropagation();
     }, []);
@@ -175,7 +168,6 @@ const PreviewView = (props: PreviewButtonProps) => {
             .then(data => {
                 if (data?.url) {
                     setUrl(data.url);
-                    // url = data.url;
                     if (
                         fileType &&
                         ['yaml', 'json', 'markdown'].indexOf(fileType) !== -1
@@ -200,6 +192,7 @@ const PreviewView = (props: PreviewButtonProps) => {
                 notify(e);
             });
     };
+
     useEffect(() => {
         handlePreview();
     }, [url]);
