+++ src/components/buttons/RowButtonGroup.tsx
@@ -2,7 +2,7 @@ import { Stack } from '@mui/material';
 import { styled } from '@mui/material/styles';
 import { ReactElement } from 'react';
 
-export const StyledButtonGroup = styled(Stack, {
+const StyledButtonGroup = styled(Stack, {
     name: 'RaRowButtonGroup',
     overridesResolver: (props, styles) => styles.root,
 })({
@@ -13,9 +13,8 @@ export const StyledButtonGroup = styled(Stack, {
 
 export const RowButtonGroup = (props: {
     children: ReactElement | ReactElement[];
-    label?: string;
 }) => {
-    const { label = 'actions', children } = props;
+    const { children } = props;
     return (
         <StyledButtonGroup direction="row" spacing={1}>
             {children}
