+++ src/resources/runs/list.tsx
@@ -17,11 +17,11 @@ import yamlExporter from '@dslab/ra-export-yaml';
 import { useState, useEffect } from 'react';
 import { FlatCard } from '../../components/FlatCard';
 import { ListPageTitle } from '../../components/PageTitle';
-import { RowButtonGroup } from '../../components/RowButtonGroup';
+import { RowButtonGroup } from '../../components/buttons/RowButtonGroup';
 import { useSchemaProvider } from '../../provider/schemaProvider';
 import { StateChips, StateColors } from '../../components/StateChips';
 import { RunIcon } from './icon';
-import { BulkDeleteAllVersions } from '../../components/BulkDeleteAllVersions';
+import { BulkDeleteAllVersionsButton } from '../../components/buttons/BulkDeleteAllVersionsButton';
 import { useRootSelector } from '@dslab/ra-root-selector';
 import { functionParser, taskParser } from '../../common/helper';
 import { ListBaseLive } from '../../components/ListBaseLive';
@@ -114,7 +114,7 @@ export const RunList = () => {
                         >
                             <Datagrid
                                 rowClick={'show'}
-                                bulkActionButtons={<BulkDeleteAllVersions />}
+                                bulkActionButtons={<BulkDeleteAllVersionsButton />}
                             >
                                 <TextField source="id" label="fields.id" />
                                 <DateField
